+++ api/src/main/java/org/openmrs/module/registrationcore/api/mpi/pixpdq/PdqPatientFetcher.java
@@ -18,6 +18,10 @@ public class PdqPatientFetcher implements MpiPatientFetcher {
     @Qualifier("registrationcore.mpiPixPdqMessageUtil")
     private PixPdqMessageUtil pixPdqMessageUtil;
 
+    @Autowired
+    @Qualifier("registrationcore.Hl7SenderHolder")
+    private Hl7SenderHolder hl7SenderHolder;
+
     protected final Log log = LogFactory.getLog(this.getClass());
 
     @Override
@@ -35,10 +39,9 @@ public class PdqPatientFetcher implements MpiPatientFetcher {
         try
         {
             Message pdqRequest = pixPdqMessageUtil.createPdqMessage(queryParams);
-            if(pixPdqMessageUtil.isHl7enabled()) {
-                Message response = pixPdqMessageUtil.getHl7v2Sender().sendPdqMessage(pdqRequest);
-                retVal = pixPdqMessageUtil.interpretPIDSegments(response).get(0);
-            }
+            Message response = hl7SenderHolder.getHl7v2Sender().sendPdqMessage(pdqRequest);
+
+            retVal = pixPdqMessageUtil.interpretPIDSegments(response).get(0);
 
             return retVal;
         }
